% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11eigen-package.R
\docType{package}
\name{cpp11eigen-package}
\alias{cpp11eigen}
\alias{cpp11eigen-package}
\title{cpp11eigen: An 'Eigen' Interface}
\description{
Provides function declarations and inline function definitions that facilitate communication between R and the 'Eigen' 'C++' library for linear algebra and scientific computing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/cpp11eigen/}
  \item \url{https://github.com/pachadotdev/cpp11eigen}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11eigen/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Gael Guennebaud (Eigen library (C++))
  \item Gael Guennebaud (Eigen library (C++))
  \item Benot Jacob (Eigen library (C++))
  \item  Intel Corporation (Eigen library (C++))
  \item  Xerox Corporation (Eigen library (C++))
}

}
\keyword{internal}
