% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{eigen_version}
\alias{eigen_version}
\title{Get eigen version}
\usage{
eigen_version()
}
\value{
A string with the eigen version name and number
}
\description{
Provides the eigen C++ library version number included in the package.
}
\examples{
eigen_version()
}
